+++ Classes/Utils/GroupLoot.lua
@@ -23,15 +23,17 @@ function GroupLoot:OnStartLootRoll(_, rollID)
 	self.Log:d("START_LOOT_ROLL", rollID)
 	if not addon.enabled then return self.Log:d("Addon disabled, ignoring group loot") end
 	local link = GetLootRollItemLink(rollID)
+	local canNeed = select(6, GetLootRollItemInfo(rollID))
 	if self:ShouldPassOnLoot() then
 		self.Log:d("Passing on loot", link)
 		self:RollOnLoot(rollID, 0)
 		self.OnLootRoll(link, rollID, 0)
 
-	elseif self:ShouldGreedOnLoot() then
-		self.Log:d("Greeding on loot", link)
-		self:RollOnLoot(rollID, 2)
-		self.OnLootRoll(link, rollID, 2)
+	elseif self:ShouldRollOnLoot() then
+		self.Log:d("Rolling on loot", link, canNeed)
+		local needGreed = canNeed or 2
+		self:RollOnLoot(rollID, needGreed)
+		self.OnLootRoll(link, rollID, needGreed)
 	end
 end
 
@@ -57,7 +59,7 @@ function GroupLoot:ShouldPassOnLoot()
 		and (db.autoGroupLootGuildGroupOnly and addon.isInGuildGroup or not db.autoGroupLootGuildGroupOnly)
 end
 
-function GroupLoot:ShouldGreedOnLoot()
+function GroupLoot:ShouldRollOnLoot()
 	return addon.mldb and addon.mldb.autoGroupLoot and addon.handleLoot and
 		addon.masterLooter and addon.isMasterLooter and GetNumGroupMembers() > 1
 end